feat: properly organize constants, schemas and types (in that order) to define leveraging the previous ones
