+++ packages/constants/src/lib/constants/constants.spec.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import {
   LIT_CHAINS,
   LIT_COSMOS_CHAINS,
@@ -14,14 +13,12 @@ describe('constants', () => {
     'xdai',
     'bsc',
     'arbitrum',
+    'arbitrumSepolia',
     'avalanche',
     'fuji',
     'harmony',
-    'kovan',
     'mumbai',
     'goerli',
-    'ropsten',
-    'rinkeby',
     'cronos',
     'optimism',
     'celo',
@@ -33,6 +30,46 @@ describe('constants', () => {
     'evmosTestnet',
     'bscTestnet',
     'baseGoerli',
+    'baseSepolia',
+    'moonbeam',
+    'moonriver',
+    'moonbaseAlpha',
+    'filecoin',
+    'hyperspace',
+    'sepolia',
+    'scrollAlphaTestnet',
+    'scroll',
+    'zksync',
+    'base',
+    'lukso',
+    'luksoTestnet',
+    'zora',
+    'zoraGoerli',
+    'zksyncTestnet',
+    'lineaGoerli',
+    'yellowstone',
+    'chiado',
+    'zkEvm',
+    'mantleTestnet',
+    'mantle',
+    'klaytn',
+    'publicGoodsNetwork',
+    'optimismGoerli',
+    'waevEclipseTestnet',
+    'waevEclipseDevnet',
+    'verifyTestnet',
+    'fuse',
+    'campNetwork',
+    'vanar',
+    'lisk',
+    'chilizMainnet',
+    'chilizTestnet',
+    'skaleTestnet',
+    'skale',
+    'fhenixHelium',
+    'hederaTestnet',
+    'bitTorrentTestnet',
+    'kintoTestnet',
   ];
 
   const MUST_HAVE_SOL_CHAINS = ['solana', 'solanaDevnet', 'solanaTestnet'];
@@ -41,6 +78,9 @@ describe('constants', () => {
     'kyve',
     'evmosCosmos',
     'evmosCosmosTestnet',
+    'cheqdMainnet',
+    'cheqdTestnet',
+    'juno',
   ];
 
   const networkPubKey =
