+++ packages/schemas/package.json
@@ -1,4 +1,4 @@
- {
+{
   "name": "@lit-protocol/schemas",
   "license": "MIT",
   "homepage": "https://github.com/Lit-Protocol/js-sdk",
